requirements.txt
pyzk
pandas==2.0.0
python-dateutil==2.8.2
tkcalendar 